+++ config/quickshell/overview-bak/shell.qml
@@ -1,6 +1,5 @@
 //@ pragma UseQApplication
 //@ pragma Env QT_QUICK_CONTROLS_STYLE=Basic
-//@ pragma Env QT_QPA_PLATFORMTHEME=
 
 import "./modules/overview/"
 import "./services/"
@@ -13,13 +12,5 @@ import Quickshell
 import Quickshell.Hyprland
 
 ShellRoot {
-    Connections {
-        target: Quickshell
-
-        function onReloadCompleted() {
-            Quickshell.inhibitReloadPopup();
-        }
-    }
-
     Overview {}
 }